Added static unit tests for register definitions
Using C11's new _Static_assert() means we can have compile time checks
for getting the register definitions correct. In this instance I take
the last register of each module and compare it to the value in the
Reference manual.

In this way we can verify that the combination of register sizes and
padding is correct up to the last register. It does *not* catch
registers or padding that are out of order or named incorrectly. Doing
so would be an order of magnitude more effort.